+++ src/components/Disclaimer.tsx
@@ -11,8 +11,12 @@ const Disclaimer = () => {
       </Text>
       and */}
       <Text
-        color="indigo"
-        onPress={() => openUrl(process.env.PRIVACY_POLICY_URL ?? "")}
+        color="purpleDark"
+        onPress={() =>
+          openUrl(
+            "https://www.privacypolicies.com/live/91cf1154-e54c-45d8-90e9-fe94a46013f6"
+          )
+        }
       >
         {/* {" "} */}
         Privacy Policy