+++ src/components/inbox/InboxList.tsx
@@ -4,10 +4,32 @@ import { Divider } from "../ListSeparator";
 import { Notification } from "../../utils/types";
 import { useData } from "../../contexts/DataContext";
 import { useCallback } from "react";
+import { Box, Text } from "../restyle";
+import { RefreshCw } from "react-native-feather";
+import { useTheme } from "../../utils/theme";
 
 const renderListRow: ListRenderItem<Notification> = ({ item }) => (
   <InboxItem notification={item} />
 );
+
+const ListFooterComponent = () => {
+  const { colors } = useTheme();
+
+  return (
+    <Box
+      flex={1}
+      flexDirection="row"
+      alignItems="center"
+      justifyContent="center"
+      margin="l"
+    >
+      <RefreshCw width={12} stroke={colors.secondary} />
+      <Text marginLeft="s" variant="caption" color="secondary">
+        Only the last 50 notifications are loaded
+      </Text>
+    </Box>
+  );
+};
 
 const InboxList = () => {
   const { notifications } = useData();
@@ -22,6 +44,7 @@ const InboxList = () => {
       ItemSeparatorComponent={Divider}
       keyExtractor={(item) => item.id}
       removeClippedSubviews
+      ListFooterComponent={ListFooterComponent}
     />
   );
 };