+++ src/screens/GetStartedScreen.tsx
@@ -1,9 +1,11 @@
 import { useLayoutEffect } from "react";
 import { ScrollView } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
+import Button from "../components/Button";
 import CopyToCliboard from "../components/CopyToCliboard";
 import { Box, Text } from "../components/restyle";
 import Toaster from "../components/Toaster";
+import { useAuth } from "../contexts/AuthContext";
 import { useData } from "../contexts/DataContext";
 import { RootStackScreenProps } from "../navigation/types";
 import { useTheme } from "../utils/theme";
@@ -13,8 +15,7 @@ type Props = RootStackScreenProps<"GetStarted">;
 const GetStartedScreen: React.FC<Props> = ({ navigation }) => {
   const { userData } = useData();
   const { colors } = useTheme();
-
-  const hasCompletedOnboarding = !userData?.onboarding;
+  const { deleteUser } = useAuth();
 
   useLayoutEffect(() => {
     if (hasCompletedOnboarding) {
@@ -35,6 +36,13 @@ const GetStartedScreen: React.FC<Props> = ({ navigation }) => {
     });
     return () => listener();
   }, [navigation, userData]);
+
+  const handleGoBackToLanding = async () => {
+    await deleteUser();
+    navigation.navigate("Landing");
+  };
+
+  const hasCompletedOnboarding = !userData?.onboarding;
 
   return (
     <SafeAreaView
@@ -80,7 +88,7 @@ const GetStartedScreen: React.FC<Props> = ({ navigation }) => {
           </BulletPointListItem>
           <CopyToCliboard
             marginTop="m"
-            content={`${userData?.hook_id}@${process.env.EMAIL_DOMAIN}`}
+            content={`${userData?.hook_id}@pfg.app`}
           />
           <Text variant="caption" color="secondary" marginTop="xs">
             Tap on this box to copy it to your clipboard!
@@ -102,10 +110,17 @@ const GetStartedScreen: React.FC<Props> = ({ navigation }) => {
             appear after confirmation).
           </BulletPointListItem>
           <BulletPointListItem>
-            Thank you for reading all the way throug! now you can procede with
+            Thank you for reading all the way throug! Now you can procede with
             the steps above and enjoy reciving all your team updates directly on
             your phone!
           </BulletPointListItem>
+          <Text variant="body" marginTop="xl">
+            In case you already have an account and went here by mistake press
+            the button below!
+          </Text>
+        </Box>
+        <Box padding="m">
+          <Button onPress={handleGoBackToLanding}>Go back to landing</Button>
         </Box>
       </ScrollView>
       <Toaster />