+++ CHANGELOG.md
@@ -1,14 +1,35 @@
 # Changelog
-## [1.0.1] - Unreleased
 
-### Added
-- When a notification is recived with the app on the foreground a toast is displayed
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+
+## [unreleased]
+
+### Changes
+
+- Only load 50 notifications in inbox
 
 ### Fixed
-- App Badge now resets on open
 
-## [1.0.0] - 2022-02-15
+- Skeleton loading bug on Inbox
+
+## [1.0.1] - 2022-05-22
+
+### Added
+
+- When a notification is received with the app in the foreground a toast is displayed
+
+### Fixed
+
+- App Badge now resets on open
+- If you are on the "Get Started" page now you can go back to the landing page
+
+## 1.0.0 - 2022-05-06
 
 ### Added
 
 - First relase on the stores
+
+[unreleased]: https://github.com/zaniluca/ping-4-gitlab/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v1.0.1